+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add `String#splitCharacters()` for getting an array of the string's characters
 * Add the `loopOverArgument()` decorator, which will loop over the given argument if it's an array
 * Add the `empty` event to FunctionQueue
+* Add the `RURL#extension` property to get/set the extension of the pathname
 
 ## 0.7.0 (2020-03-13)